% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{factorial_parallel_apply}
\alias{factorial_parallel_apply}
\title{Apply a function to all combinations of other inputs}
\usage{
factorial_parallel_apply(f, ...)
}
\arguments{
\item{f}{A function to apply.}

\item{...}{Named vector-like inputs.}
}
\value{
A data frame with column names equal to \code{names(list(...))} and
the name of \code{f}.
}
\description{
Uses \code{\link[parallel]{mclapply}}, so works where fork() can be called.
}
\examples{
test_func <- function(alpha, beta) {
  alpha^2 + beta
}

factorial_parallel_apply(test_func, alpha = 1:3, beta = 1:10)
}
