% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-definitions.R
\name{make_density_obj}
\alias{make_density_obj}
\title{make "density" object}
\usage{
make_density_obj(base_density = "norm", params = list(mean = 0, sd =
  1), sample_size = 500)
}
\arguments{
\item{base_density}{a base string for a density with d\* and r\* functions,
for example "norm", as it has "dnorm" and "rnorm".}

\item{params}{a named list of parameters shared between the d\* and r\*
functions}

\item{sample_size}{number of samples to produce from this distirbution.}
}
\value{
an S3 object of type "density", with components:
\itemize{
 \item{\emph{base}}{  string: base name of distribution.}
 \item{\emph{density}}{  function: density function of distribution, takes
   one argument, its location.}
 \item{\emph{generator}}{  function: generator function of distribution.}
}
}
\description{
make "density" object
}
