% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-definitions.R
\name{make_ratio_obj}
\alias{make_ratio_obj}
\title{make "ratio" object}
\usage{
make_ratio_obj(numerator_density, denominator_density, ratio_method,
  ratio_lambda, keep_samples = TRUE, ...)
}
\arguments{
\item{numerator_density}{a "density_sugi" object representing the density of 
the numerator.}

\item{denominator_density}{a "density_sugi" object representing the density 
of the denominator.}

\item{ratio_method}{string: either "clisf" for CLISF fitting (L1 penalty, 
no positivity constraint), or "ulsif" for ULSIF fitting (L2 penalty, 
with positivity constraint).}

\item{keep_samples}{boolean: keep the samples used for estimating the density
ratio or not.}

\item{...}{named list of other options passed to either 
\code{\link[densityratiosugiyama]{clsif}} or 
\code{\link[densityratiosugiyama]{ulsif}}}
}
\value{
an S3 object of type "ratio_sugi", with components:
\itemize{
 \item{\emph{numerator_density}}{  S3 object of type "density" for the numerator.}
 \item{\emph{denominator_density}}{  S3 object of type "density" for the 
   denominator.}
 \item{\emph{ratio_method}}{  see: ratio_method.}
 \item{\emph{ratio}}{  function: evaluates the estimated density at the requested
   location.}
 \item{\emph{basis_estimate}}{  list, basis estimate. see 
   \code{\link[densityratiosugiyama]{clsif}} and 
   \code{\link[densityratiosugiyama]{eval.basis}}.}
 \item{\emph{plot_limits}}{  list, named elements are some reasonable(?) bounds for
   plots}
}
}
\description{
estimates the density ratio for two density objects.
}
