% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-ratio.R
\name{save_ratio}
\alias{save_ratio}
\title{Save the density ratio object}
\usage{
save_ratio(ratio_obj, rds_directory)
}
\arguments{
\item{ratio_obj}{"ratio_sugi" object to be saved}

\item{rds_directory}{directory to keep rds object in}
}
\value{
whatever \code{\link[base]{saveRDS}} returns
}
\description{
Save the density ratio object
}
