% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugiyama-dr.R
\name{suigyama_dr}
\alias{suigyama_dr}
\title{Estimate a density ratio and save the outputs}
\usage{
suigyama_dr(numerator = list(base_dist = "norm", params = list(mean = 1,
  sd = 0.2)), denominator = list(base_dist = "t", params = list(df = 3,
  ncp = 0)), numerator_sample_size = 500,
  denominator_sample_size = 500, keep_samples = TRUE,
  ratio_method = "clsif", ratio_lambda = 0.1, ...,
  rds_directory = NULL, pdf_directory = NULL)
}
\arguments{
\item{numerator}{List with named components "base_dist" and "args", see 
defaults for example.}

\item{denominator}{Same as \code{numerator}.}

\item{numerator_sample_size}{Number of samples to draw from the numerator.}

\item{denominator_sample_size}{Number of samples to draw from the 
denominator.}

\item{keep_samples}{boolean, should the ratio object keep the samples used to
estimate it? FALSE saves memory, but it can be useful to keep them if the 
ratio estimator breaks in a non-obvious way.}

\item{ratio_method}{string, either "clsif" or "ulsif".}

\item{ratio_lambda}{regularisation parameter for the \code{ratio_method} 
objectives.}

\item{...}{other arugments that are passed to either 
\code{\link[densityratiosugiyama]{clsif}} or 
\code{\link[densityratiosugiyama]{ulsif}}}

\item{rds_directory}{directory to save the ratio object to as an RDS file.}

\item{pdf_diectory}{directory to save the ratio plot to.}
}
\value{
I have no idea what to return here
}
\description{
Major scaffolding function for my simulation study. The default arguments
demonstrate the expected formats for this function (and all functions within
this package). I've tried to be consistent with the naming, but there are 
some issues.
}
\details{
Important assumptions:
- The "base_dist" component of the numerator/denominator objects has to have
a corresponding dbase_dist and rbase_dist methods, which a functions of
x and n_samples respectively (and the named parameters in the args 
component).
}
